+++ libs/injective-vue/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.6](https://github.com/hyperweb-io/interchainjs/compare/injective-vue@1.17.5...injective-vue@1.17.6) (2025-09-22)
+
+**Note:** Version bump only for package injective-vue
+
+## [1.17.5](https://github.com/hyperweb-io/interchainjs/compare/injective-vue@1.17.4...injective-vue@1.17.5) (2025-09-16)
+
+**Note:** Version bump only for package injective-vue
+
+## [1.17.4](https://github.com/hyperweb-io/interchainjs/compare/injective-vue@1.17.3...injective-vue@1.17.4) (2025-09-11)
+
+**Note:** Version bump only for package injective-vue
+
 ## [1.17.3](https://github.com/hyperweb-io/interchainjs/compare/injective-vue@1.17.2...injective-vue@1.17.3) (2025-08-26)
 
 **Note:** Version bump only for package injective-vue